src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/reference/ArrayReferenceIntegerTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayReferenceIntegerTextParser extends ArrayTextParser {

    public ArrayReferenceIntegerTextParser() {
        super(Integer.class, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Integer.parseInt(element.getText());
    }

}
